Load zones using spark

+++ src/main/scala/co/thebeat/bigdata/takehomeassignment/entity/package.scala
@@ -7,16 +7,16 @@ import org.locationtech.jts.geom.Polygon
 package object entity {
   case class DriverLocation(driver: String, timestamp: java.sql.Timestamp, latitude: Double, longitude: Double)
 
-  case class AugmentedDriverLocation(driver: String, timestamp: Timestamp, id_zone: Option[Int], latitude: Double, longitude: Double)
-  case class DriverZoneSession(driver: String, session_created_at: Timestamp, id_zone: Option[Int], var count: Int)
+  case class AugmentedDriverLocation(driver: String, timestamp: Timestamp, id_zone: Option[Long], latitude: Double, longitude: Double)
+  case class DriverZoneSession(driver: String, session_created_at: Timestamp, id_zone: Option[Long], var count: Int)
 
   object DriverLocation {
     val timeStampFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
   }
 
-  case class Zone(zoneId: Int, polygon: Polygon)
+  case class Zone(zoneId: Long, polygon: Polygon)
 
   case class LatLong(lat: Double, lng: Double)
-  case class RawZone(id_zone: Int, polygon: List[LatLong])
+  case class RawZone(id_zone: Long, polygon: List[LatLong])
   case class RawZones(zones: List[RawZone])
 }
